tests/wpt/tests/html/canvas/element/filters/2d.filter.canvasFilterObject.tentative.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.filter.canvasFilterObject.tentative</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.filter.canvasFilterObject.tentative</h1>
<p class="desc">Test CanvasFilter() object</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Test CanvasFilter() object");
_addTest(function(canvas, ctx) {

  _assert(ctx.filter == 'none', "ctx.filter == 'none'");
  ctx.filter = 'blur(5px)';
  _assert(ctx.filter == 'blur(5px)', "ctx.filter == 'blur(5px)'");
  ctx.filter = new CanvasFilter({name: 'gaussianBlur', stdDeviation: 5});
  _assert(ctx.filter.toString() == '[object CanvasFilter]', "ctx.filter.toString() == '[object CanvasFilter]'");
  ctx.filter = new CanvasFilter([
      {name: 'gaussianBlur', stdDeviation: 5},
      {name: 'gaussianBlur', stdDeviation: 10}
  ]);
  _assert(ctx.filter.toString() == '[object CanvasFilter]', "ctx.filter.toString() == '[object CanvasFilter]'");
  var canvas2 = document.createElement('canvas');
  var ctx2 = canvas2.getContext('2d');
  ctx2.filter = ctx.filter;
  _assert(ctx.filter.toString() == '[object CanvasFilter]', "ctx.filter.toString() == '[object CanvasFilter]'");
  ctx.filter = 'blur(5px)';
  _assert(ctx.filter == 'blur(5px)', "ctx.filter == 'blur(5px)'");
  ctx.filter = 'none';
  _assert(ctx.filter == 'none', "ctx.filter == 'none'");
  ctx.filter = new CanvasFilter({name: 'gaussianBlur', stdDeviation: 5});
  ctx.filter = 'this string is not a filter and should do nothing';
  _assert(ctx.filter.toString() == '[object CanvasFilter]', "ctx.filter.toString() == '[object CanvasFilter]'");

});
</script>
